refactor: query key 도메인 단위 분리
