+++ src/apis/user/queryKey.ts
@@ -1,0 +1,4 @@
+export const USER_QUERY_KEY = {
+  ALL: ['USER'] as const,
+  ME: () => [...USER_QUERY_KEY.ALL, 'ME'] as const,
+} as const;
